Minimal + Example
Added a couple new things.

First is an example mod, it contains the required code to be loaded (LoadMod class and MinimalRequirement class) and also a small example mod (ExampleMod class)

The modloader had a small update in the phases.

Loading phase: this phase starts when the modloader script loads and will remain until the main menu is reached

MainMenu phase: this phase starts when the main menu is loaded and will remain until the user enters a game

InGame phase: this phase starts when the user enters a game (HudUI loads) and remains until game close or returned to main menu.